# web/src/app/core/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelReply": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          }
        },
        "required": [
          "cancelled"
        ],
        "title": "CancelReply",
        "type": "object"
      },
...
      "FilesReply": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileReply"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "FilesReply",
        "type": "object"
      },
      "GrantRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "GrantRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryReply": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobReply"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "HistoryReply",
        "type": "object"
      },
...
      "SubmitBody": {
        "description": "A request to make a song.\n\nValidated here as well as in the browser. The browser's validation is for\nthe person typing; this one is because the browser is not the only thing\nthat can post to this.",
        "properties": {
          "arrangement": {
            "anyOf": [
              {
                "maxLength": 1000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "bank": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Bank",
            "type": "string"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "mimicry": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mimicry"
          },
          "source_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "source_url",
          "bank"
        ],
        "title": "SubmitBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SongGenerator edge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "operationId": "history_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History"
      },
      "post": {
        "operationId": "submit_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "one_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One"
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel"
      }
    },
    "/jobs/{job_id}/files": {
      "get": {
        "operationId": "files_jobs__job_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilesReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Files"
      }
    },
    "/jobs/{job_id}/files/{name}": {
      "get": {
        "description": "Hand back one rendering.\n\nThe name is matched against the files the job actually produced rather\nthan joined onto a directory. Joining is how this kind of route becomes\na way to read any file on the machine: `..%2f..%2f` and a path outside\nthe output folder. Comparing against a listing cannot leave the folder,\nwhatever the caller sends, so the guard is the shape of the code rather\nthan a check that can be forgotten.",
        "operationId": "one_file_jobs__job_id__files__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One File"
      }
    },
...
    "/library/{song}/{bank}/{name}": {
      "get": {
        "description": "One rendering, to play or to save.\n\nContained by resolving the path and checking it is still inside the\noutput directory, rather than by searching the listing for a match.\nThe listing is the safer-looking option and is the wrong one here: a\nbrowser playing audio issues range requests, so streaming one song\nwould walk every rendering on the machine several times over, and\nthere are already more than a thousand of them.\n\nresolve() is what makes the check sound. It normalises `..` and\nfollows symlinks first, so the comparison is against where the path\nreally lands rather than how it was spelled.",
        "operationId": "track_library__song___bank___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "song",
            "required": true,
            "schema": {
              "title": "Song",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bank",
            "required": true,
            "schema": {
              "title": "Bank",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track"
      }
    },
    "/users": {
      "get": {
        "operationId": "list_users_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersReply"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Users"
      },
      "post": {
        "operationId": "grant_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant"
      }
    },
    "/users/{email}": {
      "delete": {
        "operationId": "revoke_users__email__delete",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "schema": {
              "title": "Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke"
      }
    }
  }
}